Fix optimization with incremental mode (#1178)
* Fix spelling

* Fix issue 993

This PR fixes the issue 993. Before this PR we could not use
optimization in incremental mode.

Now we store all the objectives in a vector of lists in the environment
of SatML. We push/pop this vector in `Satml.push/Satml.pop` and we add
new objective in the last list in `Satml.optimize`. All the objectives
to one incremental guard are given to the theory just after decide the
guard. The toplevel is a corner case: we do not store objectives at the
toplevel. Instead we give them directly to the theory in `Satml.optimize`.
Notice that `Satml.optimize` has to be used at the decision level 0.
This invariant was not ensure by an assertion.

This commit also fixes another tiny issue. If you consider the following
input file:
```smt2
(set-option :produce-models true)
(set-logic ALL)
(declare-const x Int)
(assert (<= x 10))
(maximize x)
(maximize x)
(check-sat)
(get-objectives)
```
we expect that AE will output twice `(x 10)` but before this commit it will
output it only once because we do not use a unique identifier per `maximize`
statement. I add a unique identifier and use it to to order them in the objective model. 

As I explained during the last dev meeting, I also have to ensure that
SatML decides all the incremental guards before answering `sat`.
Otherwise, it could answer `sat` before giving all the objective
functions to the theory. This case happens if we use push/pop without
`assert` at this incremental level:
```smt2
(set-option :produce-models true)
(set-logic ALL)
(declare-const x Int)
(assert (<= x 10))
(push 1)
(maximize x)
(check-sat)
(get-model)
(get-objectives)
(check-sat)
(pop 1)
```
On `next`, AE answers `x = 0` here.

I also rename the field `next_objective` to `next_optimized_split`
because it contains the last split produced by the optimization
procedure of a theory.