Do not use lock file for all OCaml versions
In the current version of our workflow, we use an opam lock file to
ensure that we can reproduce builds. Unfortunately, there are different
conflicts depending on the plateform and OCaml version. We use a hack to
prevent opam from saving the version of OCaml's toolchain (see the
target `lock` in our Makfile). This hack is not sufficient if we want to
use the current workflow `build` on Windows.

This PR modifies the workflow `build` and the Makefile as follows:
1. We check the reproducibility only on OCaml 4.14.1. I choose this
   version because we both develop on this version with Nix.
2. We checked the Makefile on OCaml 4.14.2, which is strange because we
   check the installation with opam on OCaml 4.14.1. I replace 4.14.2 by
   4.14.1.
3. I remove the flag `-y` in the target `dev-switch`, `deps` and
   `test-deps` because we use `OCAMLYES` in our workflows and Makefile's
   users want to check dependencies before running a slow task as `opam
   switch create` or `opam install`.
4. I remove the hack in the target `lock`.
   If we want to update the lock file, we have to do it in a Nix shell
   with `make lock`.