Restore Windows CI on Cygwin (#1203)
The latest version of setup-ocaml has a new support for Windows. As we plan to support officially the Windows platform, we should test it on our CI. This PR restores the CI on Windows.

Currently, we compile with Cygwin because there is no support for MSYS2 in opam-setup yet.

I add a new tag unix for tests whose the purpose is to ignore tests on non-unix platform.