Deprecate debug flags (#1204)
* Deprecate the debug flag 'sum'

After merging sum theory into adt theory, this flag has no effect
anymore. This commit deprecates it and now it has the same effect as the
flag 'adt'.

* Deprecate debug flag `typing`

This flag is unused.

* Deprecated warnings flag

The debug flag `warnings` is only used by legacy parser. We deprecate it
in `v2.6.0` and it will be removed with the legacy parser in `v2.7.0`.