fix(BV, CP): Run cross-propagators to completion (#1221)
Since #1185, when substitutions change the domain of a variable, we no
longer trigger propagations, which means that we can end up in a state
where the bitlist and interval domains are not consistent (i.e. running
[constrain_bitlist_from_interval] or [constrain_interval_from_bitlist]
would shrink some domains).

This was initially part of #1185 but was accidentally removed as part of
a simplification pass during review. Add it back.