fix(opt): Only perform optimization when building a model (#1224)
The optimization module is supposed to help us build an optimized model,
so it doesn't make much sense to start optimizing before we start
actually looking for a model.

Previously, we were kind of forced into it because it would be incorrect
to perform case splits before optimizing. Now that optimization is
integrated directly into the SAT solver, it is fully independent from
case splits, and this restriction does not apply any longer.

This fixes issues where we would try to optimize eagerly in a small
solution space and would end up enumerating the solution space before
performing examining some decisions that would prune it for us, which is
exactly what happened in #1222.

Separate optimization (`do_optimize`) from case splitting
(`do_case_split`) at the `Theory` level, and perform optimization in
`compute_concrete_model`, i.e. at the time we switch to model
generation (note: this limits the impact of optimization on unsat
problems).

Also change the order of decisions to consider optimized splits last for
consistency, although that should not have much impact in practice.