Add js_of_ocaml debug flags
Currently, we compile the js worker without debug flags in our makefile.
I got an annoying bug while adapting the worker for Dolmen and without the
dev flags of `dune`, it was impossible to debug. I think it is better to
build it with the profile `dev` by default.

I also add more debugs flags at the toplevel of our worktree.