Disable zip feature in Bytecode mode
As Basile noticed, the Javascript code in `My_zip` is dead too. I
removed it and clean a bit the module.

I also produces an appropriate `run error` if we try to use the feature
in Javascript because `camlzip` does not work properly in bytecode.

Currently, the error cannot be trigger because we do not use yet the
Dolmen version of `Solving_loop` in the javascript worker. This will
change in my next PR.