Uniformize Dolmen alias modules
Dolmen paths are a bit mouthful so we use aliases. Most of these
aliases have been introduced in `Solving_loop` but we do not always
stick to this conventions and we have to repeat them at the beginning of
many files. This commit introduces a new module `Alias` which contains
these aliases and can be open in implementation files.