Better printer + remove dead code
The function `xs_modulo_records` is not useful anymore. Indeed, all the
ADT variables are sent to the ADT relation thanks to the `init` function
of `Uf`. As record type has only one constructor by definition, this
will always replace a record variable by a record definition in the
union find and we can match patterns of the form `{ lbl1 = ..; ...; lblk =
..}`.