Fix issue 1243
The fix consists in using `declared_ids` to guarantee that we never
define symbols that have not been declared at the current assertion
level in models. A proper fix requires a lot of work as we need to
rework the push/pop mechanism of CDCL.

This PR is rebased #1280 to ensure we won't mix identifiers from
different assertion levels as it could happen with string-based
identifiers.