Add a test to cover corner cases in `add_predicate`
The Instances` does not handle ground predicates the same way as other
predicates. These corner cases are not covered by our non-regression tests
(add an assertion if you do not believe me).

This PR adds a test in AE format to cover these cases.

Note that we cannot cover them with a SMT format test, as we
do not differentiate functions and predicates in `Translate` for
this format.