packaging: Avoid Cmdliner 2.0 (#1308)
Cmdliner is planning to drop parsing of options using prefixes [1] in
version 2.0. This is a change that can potentially impact our users in
unpredicted and unwanted ways, so let's make sure we only move to
cmdliner 2.0 after deciding we are OK with dropping this feature and a
review of potential uses in Alt-Ergo. In the meantime, add 2.0 as an
upper bound for cmdliner.

[1] : https://github.com/dbuenzli/cmdliner/issues/200