Centralize handling of disjunction order in `mk_or`
Clauses have an additional `is_impl` boolean flag tracking whether they
originated from an implication. This flag is used in two places:

 - To determine the order in which the Tableaux solver decides on
 literals;

 - In the `resolution_of_disj` function used for backward/forward
 triggers (also only used by the Tableaux solver).

The logic for determining the order between disjunction operands used to
be partially in the `mk_or` smart constructor and partially in the
Tableaux solver itself. This patch moves all of that logic into the
`mk_or` smart constructor.

Note that the CDCL solver currently ignores this order and always sorts
disjunction arguments (and ignores `is_impl`) using `Expr.compare` after
flattening, but this makes it easier to bring both solvers in line in
the future.