Do not leak non-abstract names in model
This patch changes the definition of `is_model_term` to align with the
definition of a _value term_ in the SMT-LIB standard (value terms are
particular ground terms defined in a logic for each sort (see Subsection
5.5.1) in [1].)

It is now the responsibility of `Uf.assign_next` and `X.assign_value` to
ensure they create value terms for all terms in the model, which
involves assigning them new names of `abstract` kind. In particular, we
no longer need to create abstract names in the `Uf` module.

In order to make sure those names show up as the canonical
representative when they exist, they are given a depth of `0`, i.e. less
than any other term (except the `true` and `false` constants).

Fixes #1213
Fixes #1270

[1]: The SMT-LIB Standard Version 2.7
     https://smt-lib.org/papers/smt-lib-reference-v2.7-r2025-07-07.pdf