Upgrade dolmen version
Use the current development version of Dolmen, as preparation for its
next release.

This patch does the minimal amount of work needed to make Alt-Ergo
support this version of Dolmen ; in particular, it does not exploit
Dolmen's new organization of builtins by theories or implement new
operators from the SMT-LIB 2.7 standard (except for `ubv_to_int` and
`int_to_bv`, which are now new names for `bv2nat` and `int2bv`
respectively).